Voltando a usar Integer no lugar de Long

+++ FuncaoLogica.java
@@ -2,15 +2,15 @@ import java.util.*;
 
 public class FuncaoLogica {
     private int numeroDeVariaveis;
-    private Set<Long> mintermos;
-    public FuncaoLogica(int numeroDeVariaveis, Set<Long> mintermos) {
+    private Set<Integer> mintermos;
+    public FuncaoLogica(int numeroDeVariaveis, Set<Integer> mintermos) {
         this.numeroDeVariaveis = numeroDeVariaveis;
         this.mintermos = mintermos;
     }
     public int getNumeroDeVariaveis() {
         return numeroDeVariaveis;
     }
-    public Set<Long> getMintermos() {
+    public Set<Integer> getMintermos() {
         return mintermos;
     }
 }
